🐛 fix : 알림 페이지 스크롤 중첩 문제 수정
