+++ Makcha/src/pages/History/components/HistoryPastConfirmOverlay.tsx
@@ -15,13 +15,13 @@ const HistoryPastConfirmOverlay = ({ open, route, detail, onClose }: Props) => {
     if (!open || !route || !detail) return null;
 
     return (
-        <div className="absolute inset-0 z-50 flex">
+        <div className="fixed inset-0 z-50 flex">
             <div
                 className="absolute inset-0 bg-black/10 dark:bg-black/30 transition-opacity"
                 onClick={onClose}
             />
 
-            <Panel width="md:w-100" disablePadding className="shadow-2xl relative z-10">
+            <Panel width="md:w-100" disablePadding className="relative z-10 shadow-2xl">
                 <RouteConfirmPanel
                     route={route}
                     detail={detail}
@@ -31,7 +31,7 @@ const HistoryPastConfirmOverlay = ({ open, route, detail, onClose }: Props) => {
                 />
             </Panel>
 
-            <section className="hidden md:block min-w-0 flex-1 h-dvh relative z-10">
+            <section className="relative z-10 hidden min-w-0 flex-1 md:block h-full">
                 <KakaoMapView routes={[route]} selectedRouteId={route.id} />
             </section>
         </div>
